# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "realt backend api",
    "description": "realt backend",
    "version": "1.0.0"
  },
  "host": "realt-backend.herokuapp.com",
  "basePath": "",
...
  "schemes": ["https"],
  "paths": {
    "/realtors": {
      "post": {
        "tags": ["realtors"],
        "summary": "create new realtor",
        "parameters": [
          {
            "in": "body",
            "name": "realtor",
            "schema": {
              "$ref": "#/definitions/Realtor"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": ["realtors"],
        "summary": "get realtor",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": ["realtors"],
        "summary": "update realtor",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "write access token here",
            "type": "string"
          },
          {
            "in": "body",
            "name": "realtor",
            "schema": {
              "$ref": "#/definitions/RealtorUpdate"
            }
          }
        ],
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Realtor": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "passwordRepeat": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "agency": {
          "type": "boolean"
        },
        "agencyname": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "site": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password",
        "passwordRepeat",
        "agency",
        "phone",
        "name",
        "surname"
      ]
    },
    "RealtorUpdate": {
      "properties": {
        "phone": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "patronymic": {
          "type": "string"
        },
        "agency": {
          "type": "boolean"
        },
        "agencyname": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "site": {
          "type": "string"
        }
      },
      "required": ["agency", "phone", "name", "surname"]
    },
...
  }
}